Optimized Cancelations to register and unregister in O(1) time instead of O(n).

Master:

|                Method | Registrations |        Mean |     Error |    StdDev | Allocated | Survived |
|---------------------- |-------------- |------------:|----------:|----------:|----------:|---------:|
| RegisterAndUnregister |             1 |    281.0 ns |   1.38 ns |   1.29 ns |         - |     72 B |
| RegisterAndUnregister |            10 |  3,019.9 ns |   8.10 ns |   6.76 ns |         - |    320 B |
| RegisterAndUnregister |           100 | 35,129.4 ns | 144.63 ns | 128.21 ns |         - |  3,200 B |

This PR:

|                Method | Registrations |       Mean |    Error |   StdDev | Allocated | Survived |
|---------------------- |-------------- |-----------:|---------:|---------:|----------:|---------:|
| RegisterAndUnregister |             1 |   109.9 ns |  2.20 ns |  2.62 ns |         - |     56 B |
| RegisterAndUnregister |            10 |   995.9 ns |  6.82 ns |  6.38 ns |         - |    560 B |
| RegisterAndUnregister |           100 | 9,866.8 ns | 49.43 ns | 75.49 ns |         - |  5,600 B |